immigration_translator_offline/requirements.txt
# Core FastAPI and web dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# ASR
openai-whisper==20231117
torch==2.2.0

# Translation
transformers==4.35.2
sentencepiece==0.2.0
tokenizers==0.15.2
accelerate==0.25.0
safetensors==0.4.1

# TTS (Windows-friendly)
pyttsx3==2.90

# Audio processing
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.4
scipy==1.11.4

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
huggingface-hub==0.19.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2